tests work - want to think about making them headless
