+++ README.md
@@ -4,18 +4,29 @@
 ### official installation page
 https://chrome.google.com/webstore/detail/amazon-order-history-repo/mgkilgclilajckgnedgjgnfdokkgnibi
 
-Maybe the UID on the end of the URL changes when we publish a new version?
-
 ### linting (uses npm lint tools)
 ```
-./lint.sh
+npm run lint
+```
+
+### building
+```
+npm run build
+```
+
+### packaging
+```
+npm run package
 ```
 
 ### installing locally on chrome
 * Open chrome, and type chrome://extensions in the address bar.
 * Click "Load unpacked extension...".
-* Navigate to the azad folder.
+* Navigate to the build folder.
+
+### running 'unit' tests
+* build (see above)
+* open/refresh src/tests/tests.html in your browser
 
 ### edit-test-loop
-Remember to refresh/reload both on the extension page and targetted amazon page after saving changes to files.
-Test code can be exercised in the browser by opening ./tests/tests.html
+Remember to re-build, reload the extension on the chrome://extensions page and reload the targetted amazon page after saving changes to files.
