+++ src/tests/tests.html
@@ -1,17 +1,10 @@
 <!DOCTYPE html>
 <html>
     <head>
-        <script src="../sprintf.js"></script>
-        <script src="../date.js"></script>
-        <script src="../lzjs.js"></script>
-        <script src="order_data.js"></script>
-        <script src="tests.js"></script>
-        <script src="cache.test.js"></script>
-        <script src="date.test.js"></script>
-        <script src="extraction.test.js"></script>
+        <script src="../../build/alltests.bundle.js"></script>
         <link rel="stylesheet" type="text/css" href="tests.css">
     </head>
-    <body onload="tests.runAll(document)">
+    <body>
         <table id="results_table">
             <tbody id="results_table_body">
                 <tr>
